fixed 0 percent and notifications text
