+++ src/client/components/EditorExtended/components/addbutton.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { TransitionGroup, CSSTransition } from 'react-transition-group';
 import { connect } from 'react-redux';
-import { injectIntl } from "react-intl";
+import { injectIntl } from 'react-intl';
 import { getSelectedBlockNode } from '../util';
 import { getIsClearSearchObjects } from '../../../../store/searchStore/searchSelectors';
 import './addbutton.less';
@@ -148,33 +148,33 @@ export default class AddButton extends React.Component {
                 {this.props.intl.formatMessage({ id: 'insert_btn', defaultMessage: 'Insert' })}
               </div>
               <div className="act-buttons-grid">
-              {this.props.sideButtons.map(button => {
-                const Button = button.component;
-                const extraProps = button.props ? button.props : {};
+                {this.props.sideButtons.map(button => {
+                  const Button = button.component;
+                  const extraProps = button.props ? button.props : {};
 
-                return (
-                  <CSSTransition
-                    key={button.title}
-                    classNames="md-add-btn-anim"
-                    appear
-                    timeout={{
-                      enter: 200,
-                      exit: 100,
-                      appear: 100,
-                    }}
-                  >
-                    <Button
-                      {...extraProps}
-                      handleObjectSelect={this.props.handleObjectSelect}
-                      getEditorState={this.props.getEditorState}
-                      setEditorState={this.props.setEditorState}
-                      close={this.openToolbar}
-                      renderControl={this.renderControlElem}
-                      handleHashtag={this.props.handleHashtag}
-                    />
-                  </CSSTransition>
-                );
-              })}
+                  return (
+                    <CSSTransition
+                      key={button.title}
+                      classNames="md-add-btn-anim"
+                      appear
+                      timeout={{
+                        enter: 200,
+                        exit: 100,
+                        appear: 100,
+                      }}
+                    >
+                      <Button
+                        {...extraProps}
+                        handleObjectSelect={this.props.handleObjectSelect}
+                        getEditorState={this.props.getEditorState}
+                        setEditorState={this.props.setEditorState}
+                        close={this.openToolbar}
+                        renderControl={this.renderControlElem}
+                        handleHashtag={this.props.handleHashtag}
+                      />
+                    </CSSTransition>
+                  );
+                })}
               </div>
             </TransitionGroup>
           ))}
