+++ src/client/components/Story/Body.less
@@ -133,7 +133,7 @@
     }
   }
 
-  .public-DraftStyleDefault-pre pre{
+  .public-DraftStyleDefault-pre pre {
     overflow: auto;
     padding: 0;
     margin-bottom: 0;
