+++ src/client/wallet/TransfersCards/SetWithdrawVestingRoute/SetWithdrawVestingRoute.less
@@ -1,0 +1,8 @@
+@import (reference) '../../../styles/custom.less';
+@import (reference) '../../../styles/modules/variables.less';
+
+.SetWithdrawVestingRoute {
+  &__percent {
+    color: @grey-mortar;
+  }
+}
