+++ src/common/locales/default.json
@@ -1861,6 +1861,6 @@
   "withdraw_vesting_route_is_set": "Withdraw vesting route is set",
   "percent": "Percent",
   "withdraw_vesting_route_is_canceled": "Withdraw vesting route is canceled",
-  "withdraw_vesting_route_is_canceled_to": "Withdraw vesting route is canceled to {to_account}",
-  "withdraw_vesting_route_is_canceled_from": "Withdraw vesting route is canceled from {from_account}"
+  "withdraw_vesting_route_is_canceled_to": "You canceled withdraw vesting to {to_account}",
+  "withdraw_vesting_route_is_canceled_from": "{from_account} canceled withdraw vesting to you"
 }
